+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "license": "MIT",
   "scripts": {
-    "lint": "npx eslint '**/*.ts' && npx prettier --check '**/*.ts'",
+    "lint": "npx eslint '**/*.ts' && npx prettier --check '**/*.ts' && npx tsc --noEmit",
     "lint:fix": "npx eslint '**/*.ts' --fix && npx prettier --write '**/*.ts'"
   },
   "dependencies": {
